Tracks are now added properly, added some nice transition animations

+++ HulApp/app/src/main/java/pl/kamilszustak/hulapp/model/Track.kt
@@ -19,7 +19,7 @@ import java.util.*
 data class Track(
     @SerializedName("time_start") var startDate: Date = Date(),
     @SerializedName("time_end") var endDate: Date = Date(),
-    var duration: Double = 0.0,
+    var duration: Long = 0,
     @SerializedName("track_length") var length: Double = 0.0,
     @SerializedName("id_user") var userId: Int = 0
 ) : ApplicationEntity()
